coappraisal

+++ DTO_Models_For_GoodNewsGenerator/NewsModelDTO.cs
@@ -15,7 +15,7 @@ namespace DTO_Models_For_GoodNewsGenerator
 
         public virtual IEnumerable<CommentModelDTO> Comments { get; set; }
 
-        public float CoefficientPositive { get; set; }
+        public int CoefficientPositive { get; set; }
 
         public DateTime DateTime { get; set; }
